add histogram filling methods
Add methods to easily fill 1D and 2D histograms. There is now a
structure "Histo" which will infer if it is describing 1D or 2D
histograms based on the number of input parameters in its constructor.
The add_histogram methods takes in a Histo object and adds it to a
global map of loaded histograms keyed by the histograms name. A new
method "TruthSelectorBase::save_histograms(TFile*)" will save all of the
loaded histograms to the provided file. I have not added a default
storing of any histograms added.

During TSelector::Init, the user calls add_histogram(Histo) and then
during the event loop they can call fill_histo() as they need.

This may not be the smartest way, but it is a way.