add lwtnn
Now added dependency on lwtnn, which is in the release for R21.

In TruthNtupler I am loading a hardcoded NN (the one I currently have
limits for in the 2L WWbb analysis), and this should perhaps be handled
better though I don't see it changing that often.

I store the NN outputs and the log ratio discriminants of the NN
outputs. I should be handling the default variable values correctly.